estuary-cdk: move document models to document.py
This refactor will help prevent circular imports in future commits.
