+++ estuary-cdk/estuary_cdk/capture/task.py
@@ -55,7 +55,7 @@ def orjson_default(obj):
     if isinstance(obj, decimal.Decimal):
         return str(obj)
     if isinstance(obj, (bytes, bytearray)):
-        return base64.b64encode(obj).decode('utf-8')
+        return base64.b64encode(obj).decode("utf-8")
     raise TypeError
 
 
@@ -147,16 +147,21 @@ class Task:
         Or, reset() will discard any queued documents."""
 
         if isinstance(document, dict):
-            b = orjson.dumps({
-                "captured": {
-                    "binding": binding,
-                    "doc": document,
-                }
-            }, default=orjson_default)
+            b = orjson.dumps(
+                {
+                    "captured": {
+                        "binding": binding,
+                        "doc": document,
+                    }
+                },
+                default=orjson_default,
+            )
         else:
-            b = Response(
-                captured=response.Captured(binding=binding, doc=document)
-            ).model_dump_json(by_alias=True, exclude_unset=True).encode()
+            b = (
+                Response(captured=response.Captured(binding=binding, doc=document))
+                .model_dump_json(by_alias=True, exclude_unset=True)
+                .encode()
+            )
 
         self._buffer.write(b)
         self._buffer.write(b"\n")
@@ -172,9 +177,15 @@ class Task:
         """Write a SourcedSchema message for the given binding to the buffer.
         SourcedSchema messages won't be emitted until checkpoint() is called."""
 
-        b = Response(
-            sourcedSchema=response.SourcedSchema(binding=binding_index, schema_json=schema)
-        ).model_dump_json(by_alias=True, exclude_unset=True).encode()
+        b = (
+            Response(
+                sourcedSchema=response.SourcedSchema(
+                    binding=binding_index, schema_json=schema
+                )
+            )
+            .model_dump_json(by_alias=True, exclude_unset=True)
+            .encode()
+        )
 
         self._buffer.write(b)
         self._buffer.write(b"\n")
